+++ .fatherrc.ts
@@ -1,7 +1,10 @@
 export default {
-  esm: 'rollup',
-  cjs: 'rollup',
-  target: 'browser',
-  extractCSS: true,
-  injectCSS: false,
+  esm: {
+    input: 'src', // 默认编译目录
+    platform: 'browser', // 默认构建为 Browser 环境的产物
+    output: 'es',
+  },
+  umd: {
+    output: 'dist',
+  },
 };