+++ src/index.ts
@@ -1,0 +1,6 @@
+import './style';
+// import './components/button';
+// import './components/page-header';
+import Table from './table';
+
+export { Table };