+++ src/table/index.tsx
@@ -1,0 +1,11 @@
+/**
+ * 1. 代理组件 实现代码  xxx/index.jsx
+ * 2. 代理组件 样式代码  xxx/style
+ * */
+import { Table } from 'antd';
+import 'antd/es/table/style';
+
+// 自定义样式
+import './index.less';
+
+export default Table;